ionyse_contact/requirements.txt
Django==1.3.1
MySQL-python==1.2.3
South==0.7.3
distribute==0.6.24
django-classy-tags==0.3.4.1
django-floppyforms==0.4.7
django-grappelli==2.3.7
django-mptt==0.5.2
django-sekizai==0.5
mechanize==0.2.5
wsgiref==0.1.2
yolk==0.4.3
